+++ compressible_embeddings/include/resident/resident_partition_config.hpp
@@ -8,15 +8,15 @@ class ResidentPartitionConfig
 {
 public:
     ResidentPartitionConfig() = default;
-    ResidentPartitionConfig(size_t embDimBeginIncl, size_t embDimEndExcl, MemLayout memLayout);
+    ResidentPartitionConfig(int embDimBeginIncl, int embDimEndExcl, MemLayout memLayout);
 
-    __device__ __host__ size_t getEmbDimBeginIncl() const { return m_embDimBeginIncl; }
-    __device__ __host__ size_t getEmbDimEndExcl() const { return m_embDimEndExcl; }
-    __device__ __host__ size_t getEmbDim() const { return m_embDimEndExcl - m_embDimBeginIncl; }
+    __device__ __host__ int getEmbDimBeginIncl() const { return m_embDimBeginIncl; }
+    __device__ __host__ int getEmbDimEndExcl() const { return m_embDimEndExcl; }
+    __device__ __host__ int getEmbDim() const { return m_embDimEndExcl - m_embDimBeginIncl; }
     bool operator<(const ResidentPartitionConfig& other) const;
 
 private:
-    size_t m_embDimBeginIncl = 0;
-    size_t m_embDimEndExcl = 0;
+    int m_embDimBeginIncl = 0;
+    int m_embDimEndExcl = 0;
     MemLayout m_memLayout = MemLayout::ROW_MAJOR;
 };
